inventario-nft-juego/backend/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
web3==5.24.0
python-dotenv==0.19.0
pydantic==1.8.2
requests==2.26.0
python-multipart==0.0.5 